requirements.txt
asgiref==3.8.1
asttokens==3.0.0
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
Django==5.2.1
executing==2.2.0
Flask==3.1.1
Flask-Login==0.6.3
Flask-SocketIO==5.5.1
flask-profiler==1.8.1
Flask-WTF==1.2.2
Flask-SQLAlchemy==3.1.1
fonttools==4.58.0
git-filter-repo==2.47.0
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
h11==0.16.0
httplib2==0.22.0
idna==3.10
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.6
packaging==25.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.31.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.1
python-socketio==5.13.0
pytz==2025.2
pyzmq==26.4.0
requests==2.32.3
rsa==4.9.1
simple-websocket==1.1.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.5
traitlets==5.14.3
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
wcwidth==0.2.13
Werkzeug==3.1.3
wsproto==1.2.0
WTForms==3.2.1
redis==5.0.4
email_validator==2.2.0
Flask-Mail==0.9.1
celery==5.3.6
flower==2.0.1
pytest==8.4.0
gunicorn==21.2.0
waitress==3.0.0
psycopg2-binary==2.9.9
openpyxl==3.1.5
supabase==2.15.3
alembic==1.16.2
postgrest>=0.13.0
gotrue>=2.0.0
realtime>=1.0.0